Fix/layout and refactor

+++ back/src/use-cases/delete-job.ts
@@ -1,17 +1,34 @@
+import { ResourceNotFoundError } from '@/errors/resource-not-found'
+import { UnauthorizedError } from '@/errors/unauthorized'
 import { JobsRepository } from '@/repositories/jobs-repository'
+import { UsersRepository } from '@/repositories/users-repository'
 
 interface DeleteJobUseCaseRequest {
+  userId: string
   jobId: string
 }
 
 export class DeleteJobUseCase {
-  constructor(private jobsRepository: JobsRepository) {}
+  constructor(
+    private jobsRepository: JobsRepository,
+    private usersRepository: UsersRepository,
+  ) {}
 
-  async execute({ jobId }: DeleteJobUseCaseRequest) {
+  async execute({ jobId, userId }: DeleteJobUseCaseRequest) {
+    const user = await this.usersRepository.findById(userId)
+
+    if (!user) {
+      throw new ResourceNotFoundError()
+    }
+
     const job = await this.jobsRepository.findById(jobId)
 
     if (!job) {
-      throw new Error('Job not found')
+      throw new ResourceNotFoundError()
+    }
+
+    if (user?.id !== job?.userId) {
+      throw new UnauthorizedError()
     }
 
     await this.jobsRepository.delete(jobId)
